Log ServerList entries actually served to the client (#484)
* Log ServerList entries actually served to the client

Diagnostic for a live report of a server showing Down/unreachable
despite the registry looking correct server-side; nothing currently
surfaces what host/port/online state actually went out on the wire.


* fixes

* Fix game client cipher flag

* Advertise game server list status

* Handle manor list request while entering

* Release abandoned login sessions

* Reconnect login link after send failure

* Ignore skill cooldown request

* Persist initial character position

* Accept wire-safe in-game packets

* Sync live position from validate packet

* Handle in-game lifecycle packets

* Fix player info float field width

* Fix CharInfo field order

* Align EnterWorld packet order

* Fix movement decode and server list fallback

* Address game client review fixes

* Save player position during shutdown detach

---------

+++ internal/gameserver/network/serverpackets/userinfo.go
@@ -157,10 +157,10 @@ func writeUserInfo(w *wire.Writer, s UserInfoSnapshot) {
 	w.WriteInt32(0) // flying run speed: flight is not modeled
 	w.WriteInt32(0) // flying walk speed: flight is not modeled
 
-	w.WriteFloat32(1) // movement speed multiplier: no active haste/slow effect
-	w.WriteFloat32(1) // attack speed multiplier: no active haste/slow effect
-	w.WriteFloat32(float32(collisionRadius))
-	w.WriteFloat32(float32(collisionHeight))
+	w.WriteFloat64(1) // movement speed multiplier: no active haste/slow effect
+	w.WriteFloat64(1) // attack speed multiplier: no active haste/slow effect
+	w.WriteFloat64(collisionRadius)
+	w.WriteFloat64(collisionHeight)
 
 	w.WriteInt32(int32(c.HairStyle))
 	w.WriteInt32(int32(c.HairColor))
